DB
Base de datos cambio de ids a longs, conectar bd del profe, cambio hibernate db a none

+++ src/main/java/com/webdynamos/fincas/models/Solicitud.java
@@ -15,7 +15,7 @@ import lombok.*;
 public class Solicitud {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_solicitud;
+    private Long id_solicitud;
     private int duracion;
     private int estado;
